backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sqlalchemy
psycopg2
pgvector
google.ai.generativelanguage
google.generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
